third_party/WebKit/LayoutTests/fast/css/next-sibling-changed-expected.txt

This test ensures we properly invalidate style when adding children affected by direct adjacent sibling selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(testDiv).color is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
